# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookshelf API",
        "description": "API documentation for Bookshelf application",
        "contact": {
            "email": "support@test.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bookshelves": {
            "get": {
                "tags": [
                    "Bookshelf"
                ],
                "summary": "Get all bookshelves",
                "operationId": "2025f96dae889e634ad9cd0154ea5c06",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "No bookshelves found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookshelf"
                ],
                "summary": "Create a new bookshelf",
                "operationId": "1f04af4889e5b1ee1dc359ddf7847ae1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bookshelf created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/bookshelves/{id}": {
            "get": {
                "tags": [
                    "Bookshelf"
                ],
                "summary": "Get a specific bookshelf",
                "operationId": "f9c387dfbd4d0b218bc951c22a947d66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the bookshelf",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Bookshelf not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bookshelf"
                ],
                "summary": "Update a specific bookshelf",
                "operationId": "f700b386386acb6e2f8ef6ee0efd8635",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the bookshelf",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bookshelf updated successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Bookshelf not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bookshelf"
                ],
                "summary": "Delete a specific bookshelf",
                "operationId": "14b8cca46873a2d9a9deeab8ec34efdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the bookshelf",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookshelf deleted successfully"
                    },
                    "404": {
                        "description": "Bookshelf not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearerAuth",
                "bearerFormat": "JWT"
            }
        }
    }
}